\name{GeneticTools-package}
\alias{R/GeneticTools-package}
\docType{package}

\title{
  Loose Collection of Expression and Genotype Analysis Tools, including eQTL and MDR.
}

\description{
  GeneticTools is a loose collection of analysis tools for gene expression and genotype data. Currently
  it provides an eQTL and an MDR analysis. It also contains an MDR ensemble classifier.

  This is an early implementation of the package and is still under active development and further improvements
  will follow.
}

\details{
  \tabular{ll}{
    Package: \tab GeneticTools\cr
    Type: \tab Package\cr
    Version: \tab 0.2.1\cr
    Date: \tab 2013-10-14\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@uta.fi>
}

\keyword{multivariate}

