\name{GeneticTools-package}
\alias{R/GeneticTools-package}
\docType{package}

\title{
  GeneticTools
}

\description{
  A loose collection of analysis tools for gene expression and genotype data. Currently
  it provides methods for eQTL, QTL and MDR analysis. It also contains an MDR ensemble classifier.
}

\details{
  \tabular{ll}{
    Package: \tab GeneticTools\cr
    Type: \tab Package\cr
    Version: \tab 0.3.1\cr
    Date: \tab 2015-02-04\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}

\keyword{multivariate}

