\name{GeneticTools-package}
\alias{R/GeneticTools-package}
\docType{package}

\title{
  Loose Collection of Expression and Genotype Analysis Tools, including (e)QTL and MDR.
}

\description{
  GeneticTools is a loose collection of analysis tools for gene expression and genotype data. Currently
  it provides methods for eQTL, QTL and MDR analysis. It also contains an MDR ensemble classifier.
}

\details{
  \tabular{ll}{
    Package: \tab GeneticTools\cr
    Type: \tab Package\cr
    Version: \tab 0.4\cr
    Date: \tab 2014-08-13\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@uta.fi>
}

\keyword{multivariate}

